midia-validator/requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
python-multipart==0.0.9
Pillow==10.2.0
piexif==1.1.3
hachoir==3.2.0
pillow-heif==0.15.0 